Cleanup some code (SOLID) and fix bug by switching statements arround
