+++ frank-flow/src/frontend/src/app/app.component.ts
@@ -5,10 +5,7 @@ import { ModeType } from './header/modes/modeType.enum';
 import { SettingsService } from './header/settings/settings.service';
 import { Settings } from './header/settings/settings.model';
 import { SessionService } from './shared/services/session.service';
-import { FileService } from './shared/services/file.service';
 import { CurrentFileService } from './shared/services/current-file.service';
-import { FileType } from './shared/enums/file-type.enum';
-import { File } from './shared/models/file.model';
 
 @Component({
   selector: 'app-root',
@@ -24,7 +21,6 @@ export class AppComponent implements OnInit {
     private modeService: ModeService,
     private settingsService: SettingsService,
     private sessionService: SessionService,
-    private fileService: FileService,
     private currentFileService: CurrentFileService
   ) {}
 
@@ -47,29 +43,7 @@ export class AppComponent implements OnInit {
   initializeLoadLastSessionFIle(): void {
     const lastSessionFile = this.sessionService.getSessionFile();
     if (lastSessionFile) {
-      this.fetchFileAndSetToCurrent(lastSessionFile);
+      this.currentFileService.fetchFileAndSetToCurrent(lastSessionFile);
     }
-  }
-
-  fetchFileAndSetToCurrent(file: File): void {
-    this.fileService
-      .getFileFromConfiguration(file.configuration, file.path)
-      .then((result) => {
-        if (result) {
-          this.setCurrentFile(file, result);
-        }
-      });
-  }
-
-  setCurrentFile(file: File, content: string): void {
-    const currentFile = {
-      type: FileType.FILE,
-      configuration: file.configuration,
-      path: file.path,
-      xml: content,
-      saved: true,
-      flowNeedsUpdate: true,
-    };
-    this.currentFileService.setCurrentFile(currentFile);
   }
 }
